requirements.txt
google-genai==1.3.0
beautifulsoup4==4.12.3
requests==2.32.3
lxml==5.3.0
selenium==4.26.1
pandas==2.2.3
numpy==2.1.1
tenacity==9.0.0